Technologic: Remove Config Includes
These were an aspirational thing that didn't pan out. Used to get wildly
annoying issues with the server losing logging behavior when editing any
of the classes in the include list.

Since it's saving all of four or five lines of code, and is basically
just moving this Gem's implementation into a junk drawer, I'm killin it.